online-school/requirements.txt
django
djangorestframework
pillow
drf-yasg
gunicorn
dj-database-url
psycopg2-binary
whitenoise
django_heroku